project/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
deprecation==2.1.0
Django==4.1.2
django-credit-cards==0.4.1
django-crispy-forms==1.14.0
django-email-verification==0.3.1
django-encrypted-model-fields==0.6.5
django-jet==1.0.8
django-localflavor==3.1
idna==3.4
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-stdnum==1.17
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
django-filter